Move popup and redirect code
