+++ lib/msal-browser/src/interaction_client/BaseInteractionClient.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { ICrypto, INetworkModule, Logger, AuthenticationResult } from "@azure/msal-common";
+import { BrowserConfiguration } from "../config/Configuration";
+import { BrowserCacheManager } from "../cache/BrowserCacheManager";
+import { EventHandler } from "../event/EventHandler";
+import { AuthorizationUrlRequest } from "../request/AuthorizationUrlRequest";
+import { EndSessionRequest } from "../request/EndSessionRequest";
+
+export abstract class BaseInteractionClient {
+
+    protected config: BrowserConfiguration;
+    protected browserStorage: BrowserCacheManager;
+    protected browserCrypto: ICrypto;
+    protected networkClient: INetworkModule;
+    protected logger: Logger;
+    protected eventHandler: EventHandler;
+
+    constructor(config: BrowserConfiguration, storageImpl: BrowserCacheManager, browserCrypto: ICrypto, logger: Logger, eventHandler: EventHandler) {
+        this.config = config;
+        this.browserStorage = storageImpl;
+        this.browserCrypto = browserCrypto;
+        this.networkClient = this.config.system.networkClient;
+        this.logger = logger;
+        this.eventHandler = eventHandler;
+    }
+
+    abstract acquireToken(request: AuthorizationUrlRequest): Promise<AuthenticationResult|void>;
+
+    abstract logout(request: EndSessionRequest): Promise<void>;
+}
